Safari-611.1.10/LayoutTests/platform/ios-13/fast/text/international/system-language/navigator-language/navigator-language-en-US-expected.txt

FAIL navigator.language should be en-us. Was en.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
